Refactored anagram example
