+++ anagram/anagram.js
@@ -1,10 +1,10 @@
 /**
- * Compares two strings to see if they are anagrams of each other
- * An anagram is a word or phrase formed by rearranging the letters of a different word or phrase, 
- * typically using all the original letters exactly once
- * @param {string} firstSequence original word to compare to
- * @param {string} secondSequence possible anagram
- * @returns {boolean} returns true if secondSequence is an anagram of firstSequence
+ * Compares 2 chaînes pour vérifier si une est une anagramme de l'autre
+ * Une anagramme est un mot ou une expression obtenu en 
+ * permutant les lettres d'un mot ou d'une expression de départ.
+ * @param {string} firstSequence mot original
+ * @param {string} secondSequence anagramme possible
+ * @returns {boolean} true si secondSequence est une anagramme de firstSequence, false sinon
  */
 function anagram(firstSequence, secondSequence) {
   if (firstSequence.length !== secondSequence.length) {
@@ -16,18 +16,4 @@ function anagram(firstSequence, secondSequence) {
   return sortedFirstSequence === sortedSecondSequence;
 }
 
-// Refactor
-/* 
-function Anagram() {
-  function sortString(s) {
-    return Array.from(s).sort().toString();
-  }
-
-  return function anagram(firstSequence, secondSequence) {
-    return sortString(firstSequence) === sortString(secondSequence);
-  };
-}
-
-const anagram = Anagram();
-*/
 module.exports = anagram;
