+++ anagram/package.json
@@ -11,6 +11,6 @@
   "author": "Nikolay Radoev",
   "license": "CC-BY-3.0",
   "devDependencies": {
-    "jest": "28.1.1"
+    "jest": "29.6.1"
   }
 }
